+++ analysis/studyTasks/06_procedureAnalysis.R
@@ -17,7 +17,7 @@ source("analysis/private/_procedureAnalysis.R")
 
 ### Set connection Block
 # <<<
-configBlock <- "cprdGold"
+configBlock <- "cprdAurum"
 # >>>
 
 ### Provide connection details
@@ -43,6 +43,8 @@ analysisSettings <- readSettingsFile(here::here("analysis/settings/procedureAnal
 
 
 # E. Script --------------------
+
+startSnowflakeSession(con = con, executionSettings = executionSettings)
 
 executeProcedureAnalysis(con = con,
                          executionSettings = executionSettings,